+++ src/dsc-spse-frontend.ps1
@@ -8,7 +8,6 @@ configuration ConfigSpFrontend
         [Parameter(Mandatory)] [String]$SQLServerName,
         [Parameter(Mandatory)] [String]$SQLAlias,
         [Parameter(Mandatory)] [SharePointBuild] $SharePointVersion,
-        [Parameter(Mandatory)] [String]$SharePointSitesAuthority,
         [Parameter(Mandatory)] [Boolean]$EnableAnalysis,
         [Parameter(Mandatory)] [SharePointBuildInfo[]] $SharePointBits,
         [Parameter(Mandatory)] [System.Management.Automation.PSCredential]$DomainAdminCreds,
@@ -22,7 +21,6 @@ configuration ConfigSpFrontend
     Import-DscResource -ModuleName ActiveDirectoryDsc -ModuleVersion 6.7.1
     Import-DscResource -ModuleName WebAdministrationDsc -ModuleVersion 4.2.1
     Import-DscResource -ModuleName SharePointDsc -ModuleVersion 5.7.0 # Custom workaround on Add-SPDscEvent and SPInstallPrereqs
-    Import-DscResource -ModuleName DnsServerDsc -ModuleVersion 3.0.3
     Import-DscResource -ModuleName CertificateDsc -ModuleVersion 6.0.0
     Import-DscResource -ModuleName SqlServerDsc -ModuleVersion 17.5.1 # Custom workaround on SqlSecureConnection
     Import-DscResource -ModuleName cChoco -ModuleVersion 2.6.0.0    # With custom changes to implement retry on package downloads
@@ -637,71 +635,6 @@ configuration ConfigSpFrontend
             DependsOn                          = "[Script]WaitToAvoidServersJoiningFarmSimultaneously"
         }
 
-        DnsRecordCname UpdateDNSAliasSPSites {
-            Name                 = $SharePointSitesAuthority
-            ZoneName             = $DomainFQDN
-            DnsServer            = $DCServerName
-            HostNameAlias        = "$ComputerName.$DomainFQDN"
-            Ensure               = "Present"
-            PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn            = "[SPFarm]JoinSPFarm"
-        }
-
-        DnsRecordCname UpdateDNSAliasOhMy {
-            Name                 = $MySiteHostAlias
-            ZoneName             = $DomainFQDN
-            DnsServer            = $DCServerName
-            HostNameAlias        = "$ComputerName.$DomainFQDN"
-            Ensure               = "Present"
-            PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn            = "[SPFarm]JoinSPFarm"
-        }
-
-        DnsRecordCname UpdateDNSAliasHNSC1 {
-            Name                 = $HNSC1Alias
-            ZoneName             = $DomainFQDN
-            DnsServer            = $DCServerName
-            HostNameAlias        = "$ComputerName.$DomainFQDN"
-            Ensure               = "Present"
-            PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn            = "[SPFarm]JoinSPFarm"
-        }
-
-        Script WarmupSites {
-            SetScript            =
-            {
-                $jobBlock = {
-                    $uri = $args[0]
-                    try {
-                        Write-Verbose -Verbose -Message "Connecting to $uri..."
-                        # -UseDefaultCredentials: Does NTLM authN
-                        # -UseBasicParsing: Avoid exception because IE was not first launched yet
-                        # Expected traffic is HTTP 401/302/200, and $Response.StatusCode is 200
-                        Invoke-WebRequest -UseBasicParsing -Uri $uri -UseDefaultCredentials -TimeoutSec 40 -ErrorAction SilentlyContinue
-                        Write-Verbose -Verbose -Message "Connected successfully to $uri"
-                    }
-                    catch [System.Exception] {
-                        Write-Verbose -Verbose -Message "Unexpected error while connecting to '$uri': $_"
-                    }
-                    catch {
-                        # It may typically be a System.Management.Automation.ErrorRecord, which does not inherit System.Exception
-                        Write-Verbose -Verbose -Message "Unexpected error while connecting to '$uri'"
-                    }
-                }
-                [System.Management.Automation.Job[]] $jobs = @()
-                $uri = (Get-SPWebApplication)[0].Url
-                Write-Verbose -Verbose -Message "Warming up '$uri'..."
-                $jobs += Start-Job -ScriptBlock $jobBlock -ArgumentList @($uri)
-
-                # Must wait for the jobs to complete, otherwise they do not actually run
-                Receive-Job -Job $jobs -AutoRemoveJob -Wait
-            }
-            GetScript            = { return @{ "Result" = "false" } } # This block must return a hashtable. The hashtable must only contain one key Result and the value must be of type String.
-            TestScript           = { return $false } # If it returns $false, the SetScript block will run. If it returns $true, the SetScript block will not run.
-            PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn            = "[DnsRecordCname]UpdateDNSAliasSPSites"
-        }
-
         Script SetFarmPropertiesForOIDC {
             SetScript            = 
             {
@@ -892,7 +825,6 @@ $DCServerName = "DC"
 $SQLServerName = "SQL"
 $SQLAlias = "SQLAlias"
 $SharePointVersion = "SPRTM" #"SPLatest"
-$SharePointSitesAuthority = "spsites"
 $EnableAnalysis = $true
 $SharePointBits = @(
     @{
@@ -917,7 +849,7 @@ $SharePointBits = @(
 )
 
 $outputPath = "C:\Packages\Plugins\Microsoft.Powershell.DSC\2.83.5\DSCWork\ConfigureFESE.0\ConfigSpFrontend"
-ConfigSpFrontend -DomainAdminCreds $DomainAdminCreds -SPSetupCreds $SPSetupCreds -SPFarmCreds $SPFarmCreds -SPPassphraseCreds $SPPassphraseCreds -DNSServerIP $DNSServerIP -DomainFQDN $DomainFQDN -DCServerName $DCServerName -SQLServerName $SQLServerName -SQLAlias $SQLAlias -SharePointVersion $SharePointVersion -SharePointSitesAuthority $SharePointSitesAuthority -EnableAnalysis $EnableAnalysis -SharePointBits $SharePointBits -ConfigurationData @{AllNodes=@(@{ NodeName="localhost"; PSDscAllowPlainTextPassword=$true })} -OutputPath $outputPath
+ConfigSpFrontend -DomainAdminCreds $DomainAdminCreds -SPSetupCreds $SPSetupCreds -SPFarmCreds $SPFarmCreds -SPPassphraseCreds $SPPassphraseCreds -DNSServerIP $DNSServerIP -DomainFQDN $DomainFQDN -DCServerName $DCServerName -SQLServerName $SQLServerName -SQLAlias $SQLAlias -SharePointVersion $SharePointVersion -EnableAnalysis $EnableAnalysis -SharePointBits $SharePointBits -ConfigurationData @{AllNodes=@(@{ NodeName="localhost"; PSDscAllowPlainTextPassword=$true })} -OutputPath $outputPath
 Set-DscLocalConfigurationManager -Path $outputPath
 Start-DscConfiguration -Path $outputPath -Wait -Verbose -Force